Squashed commit of the following:
commit fbaac340608a52396025fada765719588834a8fb
Author: idiot <12016610+idiotbitz@users.noreply.github.com>
Date:   Wed Aug 17 04:23:57 2022 -0400

    fix for previous commit

my previous commit didn't account for when multiple tooltips were
assigned to a single menu item, causing LANGUAGE keywords to not be
fixed.

now the text is localized when the tooltip is first created, so now
there should be zero instances where LANGUAGE text isn't properly
localized in tooltips.

commit 7a386a17f8becaee02efc61f3a59ea93efba2d79
Author: idiot <12016610+idiotbitz@users.noreply.github.com>
Date:   Tue Aug 16 06:17:58 2022 -0400

    localize menu tooltip text before wrapping

this commit fixes an issue that causes the tooltip to incorrectly
show the LANGUAGE keyword if the text is longer than 1 line.