Completely overhaul AP data storage + level select menu
Data storage is now part of the StaticEventHandler rather than stored in
the player inventory, so it's unaffected by save/load. This does mean we
need to restore it from journal replay (once that's implemented) on game
startup, but we were planning to do that anyways so.

Also contains the beginnings of the remaining data package API.