Make check spawning a bit more relaxed, and add failsafes.
I was running into problems where mods like Trailblazer replaced actual items in
the map with spawners, causing AP to end up latching onto things like teleport
destinations or, in some cases, nothing at all.

The solution here is to wait up to 10 tics for the level to settle, and in that
time, look for actors that are interactable (i.e. part of the blockmap) to
replace. To deal with actors that spawn on conveyor belts and elevators, the
actor is allowed to be up to 2 units away from the check position on a single
axis as long as it hasn't moved on the other two axes.

If after this time there are some checks it couldn't find matching actors for,
it emits a warning message and drops them on the player's head.