gzap: WIP: First steps at multiworld datapack compatibility
In order for MW to work right, we need to have the item and location maps fully
initialized by the time python finishes parsing the GZDoomWorld class definition
-- doing it in generate_early() is too late.

So, we put the logic files in the apworld itself, and load them at module load
time.

A lot of work is still needed to support item/location deduplication across WADs
and whatnot, but this kinda mostly works.