gzap: move randomizer state into the playsim
This means it gets saved and loaded with the rest of the game and can be used
for multi-session singleplayer games. Loading an earlier save from the same
session will not rewind the randomizer state -- collected checks have still been
collected, etc -- but quitting the game and loading a save will give you exactly
the state from that save.