gzap: handle gzdoom restarts gracefully
By keeping the IPC buffer file at the full size, we ensure that even if gzdoom
restarts, it can still use it. By watching for truncations of the log file, we
follow the log even if it gets rewritten. And we use XON as a prompt to re-send
all the item data.