gzap: implement level order bias option
Higher numbers bias it towards following the base level order more closely. The
actual mechanic is "before it considers level N in logic, the exits for LOB% of
the preceding levels, rounded down, must be in logic". So, for example, at 25%,
it will always generate a route that has at least ⌊31×0.25⌋=7 levels in logic
before it can expect you fight the Icon of Sin.