gzap: add map rank calculation
This measures the "rank" of a map based on how far away it is from one of the
scan roots, rather than based on the order it appears in the wad. In particular
this means that E1M1 and E2M1 now have the same rank, rather than E2M1 being
considered a more advanced level than E1M8. It should also improve Wadsmoosh
behaviour.