gzap: emit blank line just before AP-MAP when scanning
Without this, wolf3d.ipk3 seems to leave a "V0" (or V1, V2, etc) without a
trailing newline in the log output buffer -- or something of that sort -- which
results in the line getting written as "V0AP-MAP", which obviously doesn't work.
I am not pleased with this workaround but I still don't understand the root
cause.