gzap: only use TotalTime, not MenuTime, for IGT
In principle the sum of these should be the "total time", but in practice the
MenuTime keeps ticking while we're in-game, so in-game time gets double-counted.
Rather than use MenuTime everywhere (since I'm not sure I trust it), we use
TotalTime and just don't include menus in the IGT anymore.