gzap: fix newly introduced crash when "start with all maps" is on
This is a bit of a hack -- at the end of logic import we synthesize a new
automap for each level. But with the changes to item registration, it would
collide with any organically-occurring automaps, causing them to be deduplicated
into (e.g.) "Automap [Allmap]" and "Automap []", and causing generation to fail
later when it tries to create an "Automap" for the player's inventory.

With this, when encountering a naturally occurring map, we just pretend it has
no typename, so it looks the same as the synthetic ones.