gzap: update checks in realtime when maps are collected
This includes a fair bit of refactoring of the Check types. The location is now
treated as the source of truth rather than the CheckPickup and CheckMarker
having their copies of bits of its state, and both the Pickup and Marker
re-evaluate their state once a second, allowing them to quickly change state
when the automap is collected.

The Pickup still keeps a copy of the checked bit, so it can be marked
un-pick-up-able as soon as you touch it.