gzap: include coordinates in tuning data
This is step 1 of changing how location names work -- we need some other way of
matching up checks when reading tuning data, and coordinates are the only
reliable way to do that. Notably, the game sends the coordinates for the
original check, not the coordinates it's at when you pick it up, so it'll still
work even if the check has moved.