gzap: major rewrite of check placement
Now, rather than waiting for the original item to spawn and then replacing it,
we spawn the check and then it tries to absorb the original item, if it can find
one. This means checks always spawn in the right place in the level even if the
original item can't be found.