gzap: internal refactoring of key system
This should be basically invisible to the end user, but it's a complete redesign
of how keys are tracked internally -- rathert than each Region just maintaining
a list of key typenames, there's a global collection of key info including which
levels each key is valid in (which may have cardinality >1). No behaviour is
changed, yet, but this lays the groundwork for things like multi-region keys or
autodetection of keys that spawn at runtime.