gzap: fixes for key handling on level entry/exit
Earlier changes (re)introduced a problem where if you pick up key Y in level P,
and then go level Q which does not have that keytype, you end up carrying around
key Y in that level. In this case it's because on level entry, CheckForNewKeys()
was identifying it as an unscanned key and producing a key entry for it.

With this change, we now check for unscanned keys only on level exit (for any
reason) or immediately after loading a savegame; we don't check when entering an
entirely new level or re-entering a previously visited level in the same hub.