gzap: grant keys via world spawn, improve key add/remove code
This makes two important changes to the key management code.

The first is just an optimization: we build separate sets of "keys you have but
shouldn't" and "keys you don't have but should", remove only the former from
your inventory, and add only the latter, rather than the previous approach of
removing all keys unconditionally and then re-adding the ones you needed.

The second is that keys are now granted to the player by spawning them in the
world and then calling CallTryPickup(). Doing this (rather than waiting for the
player to move) makes it less compatible with keys that are replaced with
spawners (vs. using replacement rules), but removes the possibility of the
player never getting the key, and it does fix compatibility with mods that
replace Doom keys with their own rather than just reskinning the Doom keys in
place -- DN3Doom is the only one I know of at present that does this, but there
are undoubtedly others.