gzap: when loading a savegame, restore item-vend counts from the save even if txn disagrees
This fixes the following issue:
- save your game;
- pick up a check that contains weapon X;
- vend it (either manually or with autovend;
- load your game.
Now the weapon is gone from your AP inventory (because the live apstate had a
higher txn), but it's also gone from your playsim inventory (because you didn't
have it in the save), and the check that contained it is gone from the world.
_OOPS_.

This addresses this problem by making the "number of copies vended" counters
from the savegame take precedence, even if the live state is otherwise
overwriting it.