gzap: fix linedef-based death exits
The underlying problem was that in persistent mode, everything is in the same
hubcluster, and you aren't allowed to exit between maps in the same hubcluster
while dead -- the linedef activation will fail when it actually invokes the
special.

We now catch exit linedef activation in WorldLinePreActivated(), and if the
player subsequently respawns, warp them out of the level.