gzap: apply weapon suppression on pickup rather than drop
This adds a new inventory force, PickupDetector, which uses HandlePickup to
apply weapon suppression on pickup. In particular this means that when playing
with slot-based suppression you can see a weapon drop, leave it, and come back
to it later. It is also more reliable than the old suppression mechanism.