gzap: fix generation crash when start_with_keys=true
This used to work because repr(key) == key.fqin(). I changed repr(key) to make
it more obvious when debugging whether I was looking at a DoomKey structure or a
string containing the FQIN and this also broke starting inventory population
from the keyset because it was incorrectly using the former rather than the
latter.