gzap: check sprite rework
This is a major rework of how check sprites are organized and loaded.

The unused red versions are removed, as are the pulsing glow versions. The
remainder, along with overlay icons by Dat Boi, are placed in patches, and
composited with the overlays as needed at load time.

All "builtin sprites" for AP purposes are now in the AP00 namespace, with the
frame index being used for different icons or items. "Icon sprites" used for
guessing icons from other games are in the AP01 namespace.

Note that at present, only the sprites for filler, progression, and unreachable
checks (AP00C, AP00P, and AP00Z) are wired up -- it doesn't know at runtime when
to apply the useful or trap bits, and I want to add a setting to display traps
differently in any case.