Add cache for compiled logic/tuning data
These are stored in pickle format in AP/gzdoom/cache. Based on testing
this results in a roughly 4x speedup in load time after the first, and
when modifying logic/tuning data only the specific wad being modified is
invalidated.

Two caches are created for each wad, one containing just the logic and
one containing logic+tuning, so that if the tuning data is modified
(which happens much more often than the logic) it can still use the
logic cache.