gzap: don't count a level exit as a level clear if nextmap is empty
It was counting a death without autosave as a level clear, because it
wasn't a level unload to load a new save. We can't just rely on checking
if the player is dead because then we break death exits, so instead we
check if there is actually a next map -- if not the level is being
unloaded to completely reinitialize the game and it doesn't count as a
win.