gzap: fix population of filler pool when using :vanilla
This happened because vanilla items were getting subtracted from the
pool without ever getting added to it, so they ended up with a very
large negative count, which resulted in the filler-to-slots ratio
calculated in create_items() being negative, which meant the filler pool
would end up filled with all the vanilla items and nothing else.

We resolve this by (a) no longer applying pool limits to vanilla items,
and (b) as a safety check, only return items with counts >0 to the
caller.