gzap: reject levelports to levels for which the player has no access
Ideally the player wouldn't be able to initiate the levelport at all,
because the linedef activation handler would catch it, *but* we still
need to worry about ACS teleports and the like. So we have a fallback
here: if the player enters a map, and it's a map that's known to AP, but
they don't have the access token for it, we kick them back to the hub as
soon as we're done map initialization.