Revert "gzap: add support for SecretTrigger"
This is a partial revert; some code that improves support for secret
sectors when SecretTriggers are present remains. Unfortunately,
SecretTriggers are not guaranteed to be located anywhere near the actual
secret they describe, so any future support for them in the rando is
going to have to implemented similar to secret sectors, with no fixed
position.