gzap: standardize on new position format
See protocol.md for the details, but the short form is: the position
format used in tuning files is now used in logic files as well, and has
been extended to support sector-based, TID-based, and event-based
virtual positions. The `virtual` bit in the apworld is now gone as a
result of this, as is all support for the legacy locations_by_name
lookup.

This does mean that older tuning files that don't include position data,
and newer ones that contain tuning for secrets, are invalidated. The
latter can easily be updated, the former probably need to be completely
remade.