gzap: don't permit fast travel to unvisited levels interior to hubs
In effect, we only permit using the level select screen to travel to a
level if (a) it's not part of a hubcluster, (b) it's one of the
entrances to the cluster, as defined at scan time, or (c) the player has
previously visited it.

This is effectively the mirror image of 10f4ce; it prevents the player
from using a levelport token to travel to places that are not in logic.