gzap: don't duplicate keys
When a cluster key subsumes the individual level keys that make it up,
it wasn't updating the original locations the key was sourced from,
which meant you would end up with both the per-level and per-cluster
versions of the key in the pool. With this change, the item at each
location is not used directly when populating the pool; instead, it's
looked up by name in the WAD's item table and the result used.