gzap: Reorganize addons to make each wad a separate apworld.
It turns out that the default AP client library -- which we use -- has a hard
limit of 16MB *after decompression* on messages from the server. Since the size
of the gzdoom datapackage scales with the total number of items and locations
across all wads we support, this means we are approaching that limit, and can
easily exceed it if the user scans a new wad of any significant size.

The root-cause fixes for this, like dynamic data packages and dynamic location
creation, require deep-seated changes to Archipelago itself, so we take this
bandaid solution of just saying that, apart from core, one wad == one apworld
and telling users to only install the ones they actually use.

Empirically, the limit seems to be between 130k and 150k locations.