gzap: make each wad a full-power apworld with its own yaml
This is a fairly quick and dirty approach, that just makes it illegal
for a DoomLogic to contain more than one DoomWad and then fishes the wad
out of the logic after logic loading. The main apworld is still needed
to provide all the supporting code, but each wad-specific apworld
provides the logic and the apworld definition.

This means that each wad now has its own yaml and its own ID space,
which is stable even if other wads are added or removed, which should
fix a lot of problems with UT.