gzap: fix client
The client needs to present a game name to the host that matches the
slot, the host won't infer it. So, we start with a placeholder name of
"GZDoom" (on startup we need a valid game name, but it doesn't have to
be the same one), and then use the `wad` field in XON to figure out
which game to actually present when connecting.

We also now refuse to connect if there is no connection to the game yet.